refactor(frontend): changed tables locations + minor improvements
