+++ frontend/src/container/tables/FavoriteTable/FavoriteTable.styl
@@ -1,4 +1,4 @@
-@import '~tracim_frontend_lib/src/css/Variable.styl'
+@import '../../../../../node_modules/tracim_frontend_lib/src/css/Variable.styl'
 
 .favoriteTable__row
   color fontColor
