+++ frontend/src/container/tables/FavoriteTable/FavoritesTable.jsx
