+++ frontend/src/container/tables/UserSpacesConfigTable/UserSpacesConfigTable.jsx
@@ -1,18 +1,18 @@
 import React from 'react'
 import PropTypes from 'prop-types'
-import { connect } from 'react-redux'
-import { translate } from 'react-i18next'
-
 import {
   TracimTable,
   spaceLeaveButtonColumn,
-  spaceRoleDropdownColumn,
-  toggleButtonColumn,
+  spaceUserRoleColumn,
   spanColumn,
+  toggleButtonColumn,
   SORT_BY
 } from 'tracim_frontend_lib'
 
-const AdminSpacesUserConfigLeaveTable = (props) => {
+import { connect } from 'react-redux'
+import { translate } from 'react-i18next'
+
+const UserSpacesConfigTable = (props) => {
   const isMailChecked = (data) => data.member.doNotify
 
   const onMailChange = (e, data) => {
@@ -32,21 +32,20 @@ const AdminSpacesUserConfigLeaveTable = (props) => {
       className: 'tracimTable__styles__flex__1'
     }, row => row.label, 'spaceName', SORT_BY.LABEL),
 
-    spaceRoleDropdownColumn({
+    spaceUserRoleColumn({
       header: props.t('Role'),
       tooltip: props.t('Sort by role'),
-      className: 'tracimTable__styles__flex__2'
-    }, props.onClickChangeRole),
+      className: 'tracimTable__styles__flex__1'
+    }),
 
     spaceLeaveButtonColumn({
-      className: 'tracimTable__styles__flex__1'
-    }, props.onLeaveSpaceClick, true, props.system, props.onlyManager)
+      className: ' tracimTable__styles__flex__1'
+    }, props.onLeaveSpaceClick, props.admin, props.system, props.onlyManager)
   ]
 
   if (props.system.config.email_notification_activated) {
     columns.splice(2, 0, toggleButtonColumn({
-      header: props.t('Email notif.'),
-      tooltip: props.t('Email notifications'),
+      header: props.t('Email notifications'),
       className: 'tracimTable__styles__flex__1'
     }, 'mailNotification', onMailChange, isMailChecked))
   }
@@ -55,7 +54,9 @@ const AdminSpacesUserConfigLeaveTable = (props) => {
     <TracimTable
       columns={columns}
       data={props.spaceList}
-      emptyMessage={props.t('This user is not a member of any space yet')}
+      emptyMessage={props.admin
+        ? props.t('This user is not a member of any space yet')
+        : props.t('You are not a member of any space yet')}
       filterable
       sortable
       colored
@@ -64,14 +65,14 @@ const AdminSpacesUserConfigLeaveTable = (props) => {
   )
 }
 
-AdminSpacesUserConfigLeaveTable.propsType = {
+UserSpacesConfigTable.propsType = {
   spaceList: PropTypes.array.isRequired,
   onLeaveSpaceClick: PropTypes.func.isRequired,
   onChangeSubscriptionNotif: PropTypes.func.isRequired,
-  onClickChangeRole: PropTypes.func.isRequired,
+  admin: PropTypes.bool.isRequired,
   onlyManager: PropTypes.func.isRequired
 }
 
 const mapStateToProps = ({ system }) => ({ system })
 
-export default connect(mapStateToProps)(translate()(AdminSpacesUserConfigLeaveTable))
+export default connect(mapStateToProps)(translate()(UserSpacesConfigTable))
